+++ config/params/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "config.go",
         "config_utils_develop.go",  # keep
         "config_utils_prod.go",
-        "errors.go",
         "io_config.go",
         "loader.go",
         "mainnet_config.go",
@@ -57,7 +56,6 @@ go_test(
         "//io/file:go_default_library",
         "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
-        "@com_github_prysmaticlabs_eth2_types//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@in_gopkg_yaml_v2//:go_default_library",
         "@io_bazel_rules_go//go/tools/bazel:go_default_library",
