+++ network/forks/BUILD.bazel
@@ -2,11 +2,16 @@ load("@prysm//tools/go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["fork.go"],
+    srcs = [
+        "errors.go",
+        "fork.go",
+        "ordered.go",
+    ],
     importpath = "github.com/prysmaticlabs/prysm/network/forks",
     visibility = ["//visibility:public"],
     deps = [
         "//beacon-chain/core/signing:go_default_library",
+        "//config/fieldparams:go_default_library",
         "//config/params:go_default_library",
         "//encoding/bytesutil:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
@@ -18,13 +23,17 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["fork_test.go"],
+    srcs = [
+        "fork_test.go",
+        "ordered_test.go",
+    ],
     embed = [":go_default_library"],
     deps = [
         "//beacon-chain/core/signing:go_default_library",
         "//config/params:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
         "//testing/assert:go_default_library",
+        "//testing/require:go_default_library",
         "@com_github_prysmaticlabs_eth2_types//:go_default_library",
     ],
 )
